Bump U-boot to use a 128 KB environment and use it
** This is a breaking change. **

It is not possible to upgrade firmware made with previous systems to
this one without some effort. This requires you to reflash MicroSD cards
on your BBBs.

The problem is that we need to use the U-boot environment to configure
the new device tree overlay settings and the existing 8 KB size is
insufficient.